pyproject.toml
[project]
name = "odin-be-template"
version = "0.1.0"
description = "Template for ODIN backend using uv and Django"
readme = "README.MD"
requires-python = ">=3.12"
dependencies = [
    "django",
    "djangorestframework",
    "celery",
    "django-split-settings",
    "psycopg2-binary",
    "django-environ",
    "gunicorn",
]

[tool.uv]
